Add workaround for Microsoft bug to groot job
We have to remove all of the containers before pulling down a new image.
This is a workaround for a Microsoft bug. This workaround was originally
in the windows2016fs job of the windows2016fs release but will now live
here as this is the job that is doing the unpacking of the image layers.

[#154731947]

Signed-off-by: Amin Jamali <ajamali@pivotal.io>